Use imports in mocha tests
