+++ tests/mocha/.eslintrc.json
@@ -5,50 +5,9 @@
         "mocha": true
     },
     "globals": {
-        "chai": false,
-        "sinon": false,
-        "addBlockTypeToCleanup": true,
-        "addMessageToCleanup": true,
-        "assertArrayEquals": true,
-        "assertCallBlockStructure": true,
-        "assertDefBlockStructure": true,
-        "assertDeprecationWarningCall": true,
-        "assertEventEquals": true,
-        "assertEventFired": true,
-        "assertEventNotFired": true,
-        "assertNthCallEventArgEquals": true,
-        "assertSingleDeprecationWarningCall": true,
-        "assertVariableValues": true,
-        "assertNoWarnings": true,
-        "assertWarnings": true,
-        "captureWarnings": false,
-        "createDeprecationWarningStub": true,
-        "createKeyDownEvent": true,
-        "createRenderedBlock": true,
-        "createTestBlock": true,
-        "defineBasicBlockWithField": true,
-        "defineRowBlock": true,
-        "defineStackBlock": true,
-        "defineStatementBlock": true,
-        "dispatchPointerEvent": true,
-        "createFireChangeListenerSpy": true,
-        "createGenUidStubWithReturns": true,
-        "getBasicToolbox": true,
-        "getCategoryJSON": true,
-        "getChildItem": true,
-        "getCollapsibleItem": true,
-        "getDeeplyNestedJSON": true,
-        "getInjectedToolbox": true,
-        "getNonCollapsibleItem": true,
-        "getSeparator": true,
-        "getSimpleJSON": true,
-        "getXmlArray": true,
-        "sharedTestSetup": true,
-        "sharedTestTeardown": true,
-        "simulateClick": true,
-        "testAWorkspace": true,
-        "testHelpers": true,
-        "workspaceTeardown": true
+        "chai": "readonly",
+        "sinon": "readonly",
+        "testHelpers": "readonly"
     },
     "rules": {
         "no-unused-vars": ["off"],
