+++ tests/mocha/all_tests.js
@@ -1,0 +1,58 @@
+/**
+ * @license
+ * Copyright 2020 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import './astnode_test.js';
+import './block_json_test.js';
+import './block_test.js';
+import './comment_test.js';
+import './connection_checker_test.js';
+import './connection_db_test.js';
+import './connection_test.js';
+import './contextmenu_items_test.js';
+import './cursor_test.js';
+import './dropdowndiv_test.js';
+import './event_test.js';
+import './extensions_test.js';
+import './field_angle_test.js';
+import './field_checkbox_test.js';
+import './field_colour_test.js';
+import './field_dropdown_test.js';
+import './field_image_test.js';
+import './field_label_serializable_test.js';
+import './field_label_test.js';
+import './field_multilineinput_test.js';
+import './field_number_test.js';
+import './field_registry_test.js';
+import './field_test.js';
+import './field_textinput_test.js';
+import './field_variable_test.js';
+import './flyout_test.js';
+import './generator_test.js';
+import './gesture_test.js';
+import './input_test.js';
+import './insertion_marker_test.js';
+import './json_test.js';
+import './keydown_test.js';
+import './logic_ternary_test.js';
+import './metrics_test.js';
+import './names_test.js';
+import './procedures_test.js';
+import './registry_test.js';
+import './shortcut_registry_test.js';
+import './theme_test.js';
+import './toolbox_test.js';
+import './tooltip_test.js';
+import './trashcan_test.js';
+import './utils_test.js';
+import './variable_map_test.js';
+import './variable_model_test.js';
+import './variables_test.js';
+import './widget_div_test.js';
+import './workspace_comment_test.js';
+import './workspace_svg_test.js';
+import './workspace_test.js';
+import './xml_test.js';
+import './zoom_controls_test.js';
