+++ tests/mocha/astnode_test.js
@@ -4,6 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import * as coreTestHelpers from './test_helpers.js';
+
+const {sharedTestSetup, sharedTestTeardown, workspaceTeardown} = coreTestHelpers;
+
 suite('ASTNode', function() {
   setup(function() {
     sharedTestSetup.call(this);
