+++ tests/mocha/connection_test.js
@@ -3,6 +3,12 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {
+  assertSingleDeprecationWarningCall, createDeprecationWarningStub,
+  sharedTestSetup, sharedTestTeardown, workspaceTeardown
+} = coreTestHelpers;
 
 suite('Connection', function() {
   setup(function() {
