+++ tests/mocha/field_angle_test.js
@@ -3,6 +3,11 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {
+  createTestBlock, sharedTestSetup, sharedTestTeardown
+} = coreTestHelpers;
 
 suite('Angle Fields', function() {
   setup(function() {
