+++ tests/mocha/field_dropdown_test.js
@@ -3,6 +3,11 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {
+  createTestBlock, sharedTestSetup, sharedTestTeardown
+} = coreTestHelpers;
 
 suite('Dropdown Fields', function() {
   setup(function() {
