+++ tests/mocha/field_number_test.js
@@ -3,6 +3,9 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {sharedTestSetup, sharedTestTeardown} = coreTestHelpers;
 
 suite('Number Fields', function() {
   setup(function() {
