+++ tests/mocha/field_test.js
@@ -3,6 +3,12 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {
+  addBlockTypeToCleanup, addMessageToCleanup, createDeprecationWarningStub,
+  sharedTestSetup, sharedTestTeardown, workspaceTeardown
+} = coreTestHelpers;
 
 suite('Abstract Fields', function() {
   setup(function() {
