+++ tests/mocha/index.html
@@ -25,6 +25,8 @@
     <script src="../../blocks/math.js"></script>
     <script src="../../blocks/procedures.js"></script>
     <script src="../../blocks/text.js"></script>
+
+    <script src="../../core/workspace_comment.js"></script>
   </head>
   <style>
     #blocklyDiv {
@@ -35,7 +37,6 @@
     }
   </style>
   <body>
-
     <div id="mocha"></div>
     <div id="failureCount" style="display:none" tests_failed="unset"></div>
     <script src="https://unpkg.com/chai/chai.js"></script>
@@ -47,63 +48,6 @@
         ui: 'tdd'
       });
     </script>
-
-    <script src="test_helpers.js"></script>
-    <script src="toolbox_helper.js"></script>
-
-    <script src="astnode_test.js"></script>
-    <script src="block_json_test.js"></script>
-    <script src="block_test.js"></script>
-    <script src="comment_test.js"></script>
-    <script src="connection_db_test.js"></script>
-    <script src="connection_checker_test.js"></script>
-    <script src="connection_test.js"></script>
-    <script src="contextmenu_items_test.js"></script>
-    <script src="cursor_test.js"></script>
-    <script src="dropdowndiv_test.js"></script>
-    <script src="event_test.js"></script>
-    <script src="extensions_test.js"></script>
-    <script src="field_test.js"></script>
-    <script src="field_angle_test.js"></script>
-    <script src="field_checkbox_test.js"></script>
-    <script src="field_colour_test.js"></script>
-    <script src="field_dropdown_test.js"></script>
-    <script src="field_image_test.js"></script>
-    <script src="field_label_test.js"></script>
-    <script src="field_label_serializable_test.js"></script>
-    <script src="field_multilineinput_test.js"></script>
-    <script src="field_number_test.js"></script>
-    <script src="field_registry_test.js"></script>
-    <script src="field_textinput_test.js"></script>
-    <script src="field_variable_test.js"></script>
-    <script src="flyout_test.js"></script>
-    <script src="generator_test.js"></script>
-    <script src="gesture_test.js"></script>
-    <script src="input_test.js"></script>
-    <script src="insertion_marker_test.js"></script>
-    <script src="json_test.js"></script>
-    <script src="shortcut_registry_test.js"></script>
-    <script src="keydown_test.js"></script>
-    <script src="logic_ternary_test.js"></script>
-    <script src="metrics_test.js"></script>
-    <script src="names_test.js"></script>
-    <script src="procedures_test_helpers.js"></script>
-    <script src="procedures_test.js"></script>
-    <script src="registry_test.js"></script>
-    <script src="theme_test.js"></script>
-    <script src="toolbox_test.js"></script>
-    <script src="tooltip_test.js"></script>
-    <script src="trashcan_test.js"></script>
-    <script src="utils_test.js"></script>
-    <script src="variable_map_test.js"></script>
-    <script src="variable_model_test.js"></script>
-    <script src="variables_test.js"></script>
-    <script src="widget_div_test.js"></script>
-    <script src="workspace_test.js"></script>
-    <script src="workspace_svg_test.js"></script>
-    <script src="workspace_comment_test.js"></script>
-    <script src="xml_test.js"></script>
-    <script src="zoom_controls_test.js"></script>
 
     <div id="blocklyDiv"></div>
     <xml xmlns="https://developers.google.com/blockly/xml" id="toolbox-simple" style="display: none">
@@ -149,7 +93,9 @@
       <block type="test_field_block"></block>
     </xml>
 
-    <script>
+    <script type="module">
+      import './all_tests.js';
+
       mocha.run(function(failures) {
         var failureDiv = document.getElementById('failureCount');
         failureDiv.setAttribute('tests_failed', failures);
