+++ tests/mocha/registry_test.js
@@ -8,7 +8,9 @@
  * @fileoverview Tests for Blockly.registry
  * @author aschmiedt@google.com (Abby Schmiedt)
  */
-'use strict';
+import * as coreTestHelpers from './test_helpers.js';
+
+const {sharedTestSetup, sharedTestTeardown} = coreTestHelpers;
 
 suite('Registry', function() {
   var TestClass = function() {};
