+++ tests/mocha/trashcan_test.js
@@ -3,6 +3,12 @@
  * Copyright 2019 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
+import * as coreTestHelpers from './test_helpers.js';
+
+const {
+  assertEventFired, assertEventNotFired, sharedTestSetup, sharedTestTeardown,
+  simulateClick
+} = coreTestHelpers;
 
 suite("Trashcan", function() {
   function fireDeleteEvent(workspace, xmlString) {
